requirements-dev.txt
pytest==6.2.5
Pillow==10.4.0
boto3==1.34.144
botocore==1.34.144
